+++ src/ctv-config.ts
@@ -1,9 +1,7 @@
 import * as path from 'path';
-import * as fs from 'fs';
 import * as vscode from 'vscode';
 import { sync as globSync } from 'glob';
 import { quote, normalizePath, isNodeExecuteAbleFile } from './utils';
-import minimatch from 'minimatch';
 
 export default class CtvConfig {
 	private workspaceRootPath?: string;
@@ -22,13 +20,6 @@ export default class CtvConfig {
 	 */
 	public get configFile(): string | undefined {
 		return this.getStringSetting('cucumber-tsflow.configFile');
-	}
-
-	/**
-	 * Cucumber profile to use when running or debugging tests
-	 */
-	public get profile(): string {
-		return this.getStringSetting('cucumber-tsflow.profile') ?? 'default';
 	}
 
 	/**
@@ -89,30 +80,6 @@ export default class CtvConfig {
 	}
 
 	/**
-	 * Glob pattern used to find step files in your workspace. The default supports multiple test projects in a workspace.
-	 */
-	public get stepsSelector(): string {
-		const glob = this.getStringSetting('cucumber-tsflow.stepsSelector') ?? '**/{steps,step_definitions}/**/*.ts';
-		return normalizePath(glob as string);
-	}
-
-	/**
-	 * Glob pattern used to find feature files in your workspace. The default supports multiple test projects in a workspace.
-	 */
-	public get featuresSelector(): string {
-		const glob = this.getStringSetting('cucumber-tsflow.featuresSelector') ?? '**/features/**/*.feature';
-		return normalizePath(glob as string);
-	}
-
-	/**
-	 * Max folder depth for your workspace.
-	 * Used to search parent folders for closest matching Cucumber settings file for a step file
-	 */
-	public get maxFolderDepth(): number {
-		return vscode.workspace.getConfiguration().get('cucumber-tsflow.maxFolderDepth') ?? 10;
-	}
-
-	/**
 	 * Preserve focus on editor when running tests
 	 */
 	public get preserveEditorFocus(): boolean {
@@ -132,9 +99,6 @@ export default class CtvConfig {
 			args.push('-c');
 			args.push(this.configFile);
 		}
-		// add the profile
-		args.push('-p');
-		args.push(this.profile);
 
 		const setOptions = new Set(options);
 		if (this.runOptions) {
@@ -214,92 +178,4 @@ export default class CtvConfig {
 		}
 		return this.workspaceRootPath;
 	}
-
-	/**
-	 * Gets the path where cucumber.js is found relative to the current
-	 * working file or root if no file open. Supports mono repo projects
-	 */
-	// public get currentCucumberSettingsPath(): string | undefined {
-	// 	if (this.cucumberSettingsPathFromFile) {
-	// 		return this.cucumberSettingsPathFromFile;
-	// 	}
-	// 	return this.cucumberSettingsPathFromRoot;
-	// }
-
-	/**
-	 * Helper used find the first settings file, starting from workspace root, that contains the
-	 * Profile specified in 'cucumber-tsflow.profile', which equals "default" by default.
-	 */
-	// private get cucumberSettingsPathFromRoot(): string | undefined {
-	// 	if (this.currentWorkspaceRootPath) {
-	// 		if (!this.cucumberSettingsFromRoot) {
-	// 			const cucumberFiles = globSync('**/cucumber.*', { cwd: this.currentWorkspaceRootPath });
-	// 			for (let idx = 0; idx < cucumberFiles.length; idx++) {
-	// 				const cucumberFile = cucumberFiles[idx];
-	// 				if (cucumberFile.indexOf('node_modules') < 0) {
-	// 					const settingsPath = path.join(this.currentWorkspaceRootPath!, cucumberFile);
-	// 					const data = fs.readFileSync(settingsPath);
-	// 					if (data.includes(this.profile)) {
-	// 						const cucumberIdx = cucumberFile.lastIndexOf('cucumber');
-	// 						const cucumberPath = cucumberIdx > 0 ? cucumberFile.substring(0, cucumberIdx - 1) : '';
-	// 						this.cucumberSettingsFromRoot = normalizePath(path.join(this.currentWorkspaceRootPath!, cucumberPath));
-	// 					}
-	// 				}
-	// 			}
-	// 		}
-	// 		return this.cucumberSettingsFromRoot;
-	// 	}
-	// 	return undefined;
-	// }
-
-	/**
-	 * Helper used to get the closest matching cumber settings file that contains the
-	 * Profile specified in 'cucumber-tsflow.profile', which equals "default" by default.
-	 * This will check to see if you have a step file open by checking against the
-	 * 'cucumber-tsflow.stepsSelector' glob pattern.
-	 *
-	 * If a step file is open this will search up the file hirerachy to find a closest
-	 * settings file. This will also return the previous match if checking the same folder.
-	 *
-	 * If a step file is not open this will return undefined.
-	 */
-	// private get cucumberSettingsPathFromFile(): string | undefined {
-	// 	const editor = vscode.window.activeTextEditor;
-	// 	if (editor) {
-	// 		let currentFolderPath: string | undefined = path.dirname(editor.document.fileName);
-	// 		if (
-	// 			currentFolderPath &&
-	// 			currentFolderPath !== this.currentStepFilePath &&
-	// 			minimatch(`${currentFolderPath}/test.ts`, this.stepsSelector)
-	// 		) {
-	// 			const stepFilepath = currentFolderPath;
-	// 			const maxDepth = this.maxFolderDepth;
-	// 			for (let x = 0; x < maxDepth; x++) {
-	// 				const cucumberFiles = globSync('./cucumber.*', { cwd: currentFolderPath });
-	// 				if (cucumberFiles.length > 0) {
-	// 					for (let idx = 0; idx < cucumberFiles.length; idx++) {
-	// 						const cucumberFile = cucumberFiles[idx];
-	// 						if (cucumberFile.indexOf('node_modules') < 0) {
-	// 							const settingsPath = path.join(currentFolderPath, cucumberFile);
-	// 							const data = fs.readFileSync(settingsPath);
-	// 							if (data.includes(this.profile)) {
-	// 								this.currentStepFilePath = stepFilepath;
-	// 								this.cucumberSettingsFromFile = normalizePath(currentFolderPath);
-	// 								return this.cucumberSettingsFromFile;
-	// 							}
-	// 						}
-	// 					}
-	// 				}
-	// 				currentFolderPath = path.join(currentFolderPath, '..');
-	// 			}
-	// 		} else if (
-	// 			currentFolderPath &&
-	// 			currentFolderPath === this.currentStepFilePath &&
-	// 			minimatch(`${currentFolderPath}/test.ts`, this.stepsSelector)
-	// 		) {
-	// 			return this.cucumberSettingsFromFile;
-	// 		}
-	// 	}
-	// 	return undefined;
-	// }
 }