+++ src/cucumber/cucumber-test-features.ts
@@ -6,27 +6,18 @@ import { ParsedFeature, ParsedScenario, TestFeatureStep } from '../types';
 import { normalizePath, toKebabCase } from '../utils';
 import { scanTestOutput } from './test-output-scanner';
 import { sortBy, compose, toLower, prop } from 'ramda';
-import CtvConfig from '../ctv-config';
-import useCtvConfig from '../use-ctv-config';
+import { hasMatchingTags } from '@lynxwall/cucumber-tsflow/lib/cucumber/utils';
 
 const sortByTestLabel = sortBy<vscode.TestItem>(compose(toLower, prop('label')));
 
-type ProjectFeature = {
-	projectItem: TestItem;
-	featureTestItems: Array<TestItem>;
-};
-
 export default class CucumberTestFeatures {
 	private scenarioData = new WeakMap<vscode.TestItem, ParsedScenario>();
-	private projectFeatures = new Array<ProjectFeature>();
 	private testItems = new Array<TestItem>();
 	private stepFileManager: StepFileManager;
 	private controller: vscode.TestController;
 	private cucumberTestRunner: CucumberTestRunner;
-	private ctvConfig: CtvConfig;
 
 	constructor(stepFileManager: StepFileManager, controller: vscode.TestController) {
-		this.ctvConfig = useCtvConfig().getConfig();
 		this.stepFileManager = stepFileManager;
 		this.controller = controller;
 		this.cucumberTestRunner = new CucumberTestRunner();
@@ -42,38 +33,23 @@ export default class CucumberTestFeatures {
 			}
 			const features = await this.stepFileManager.getParsedFeatures();
 			this.scenarioData = new WeakMap<vscode.TestItem, ParsedScenario>();
-
 			await Promise.all(features.map(feature => this.loadFeature(feature)));
-
-			if (this.projectFeatures.length > 1) {
-				for (let idx = 0; idx < this.projectFeatures.length; idx++) {
-					const projTestItem = this.projectFeatures[idx];
-					const sortedFeatures = sortByTestLabel(projTestItem.featureTestItems);
-					projTestItem.projectItem.children.replace(sortedFeatures);
-					this.testItems.push(projTestItem.projectItem);
-				}
-			} else if (this.projectFeatures.length === 1) {
-				const sortedFeatures = sortByTestLabel(this.projectFeatures[0].featureTestItems);
-				this.testItems.push(...sortedFeatures);
-			}
 		}
 		return this.testItems;
 	};
 
 	private async loadFeature(feature: ParsedFeature) {
-		const projectId = toKebabCase(feature.options.projectName);
-		let projectFeature = this.projectFeatures.find(x => x.projectItem.id === projectId);
-		if (!projectFeature) {
-			projectFeature = {
-				projectItem: this.controller.createTestItem(projectId, feature.options.projectName, undefined),
-				featureTestItems: new Array<TestItem>()
-			} as ProjectFeature;
-			this.projectFeatures.push(projectFeature);
-		}
 		const featureUri = vscode.Uri.file(normalizePath(feature.featureFile));
 		const featureId = toKebabCase(feature.title);
 
-		if (projectFeature.featureTestItems.find(x => x.id === featureId) === undefined) {
+		// Check to see if we should bypass this feature. Uses
+		// Tags setting from the default profile if found
+		const tagPattern = this.stepFileManager.projectTagPattern;
+		if (tagPattern && !hasMatchingTags(tagPattern, feature.tags)) {
+			return;
+		}
+		// feature is valid for this profile ... see if we already processed it
+		if (this.testItems.find(x => x.id === featureId) === undefined) {
 			const item = this.controller.createTestItem(featureId, feature.title, featureUri);
 			const scenarioItems = new Array<TestItem>();
 			if (feature.scenarios.length > 0) {
@@ -102,7 +78,7 @@ export default class CucumberTestFeatures {
 			}
 			item.children.replace(scenarioItems);
 
-			projectFeature.featureTestItems?.push(item);
+			this.testItems.push(item);
 		}
 	}
 
@@ -114,40 +90,37 @@ export default class CucumberTestFeatures {
 	public async updateTests(uri: vscode.Uri): Promise<void> {
 		const parsedFeature = await this.stepFileManager.getParsedFeature(uri);
 		if (parsedFeature && (parsedFeature.scenarios.length > 0 || parsedFeature.scenarioOutlines.length > 0)) {
-			let projectFeature = await this.getProjectFeature(parsedFeature);
-			if (projectFeature) {
-				const featureId = toKebabCase(parsedFeature.title);
-				const featureUri = vscode.Uri.file(normalizePath(parsedFeature.featureFile));
-				let testFeature = projectFeature.featureTestItems.find(x => x.id === featureId);
-				if (!testFeature) {
-					testFeature = this.controller.createTestItem(featureId, parsedFeature.title, featureUri);
-					projectFeature.featureTestItems.push(testFeature);
-				}
-				const scenarioItems = new Array<TestItem>();
-				for (let sIdx = 0; sIdx < parsedFeature.scenarios.length; sIdx++) {
-					const scenario = parsedFeature.scenarios[sIdx];
-					const testItem = this.controller.createTestItem(toKebabCase(scenario.title), scenario.title, featureUri);
-					this.scenarioData.set(testItem, scenario);
-					scenarioItems.push(testItem);
-				}
-				if (parsedFeature.scenarioOutlines.length > 0) {
-					for (let soIdx = 0; soIdx < parsedFeature.scenarioOutlines.length; soIdx++) {
-						const scenarioOutline = parsedFeature.scenarioOutlines[soIdx];
-						if (scenarioOutline.exampleScenarios.length > 0) {
-							const scenario = scenarioOutline.exampleScenarios[0];
-							scenario.lineNumber = scenarioOutline.lineNumber;
-							const testItem = this.controller.createTestItem(
-								toKebabCase(scenarioOutline.title),
-								scenarioOutline.title,
-								featureUri
-							);
-							this.scenarioData.set(testItem, scenario);
-							scenarioItems.push(testItem);
-						}
+			const featureId = toKebabCase(parsedFeature.title);
+			const featureUri = vscode.Uri.file(normalizePath(parsedFeature.featureFile));
+			let testFeature = this.testItems.find(x => x.id === featureId);
+			if (!testFeature) {
+				testFeature = this.controller.createTestItem(featureId, parsedFeature.title, featureUri);
+				this.testItems.push(testFeature);
+			}
+			const scenarioItems = new Array<TestItem>();
+			for (let sIdx = 0; sIdx < parsedFeature.scenarios.length; sIdx++) {
+				const scenario = parsedFeature.scenarios[sIdx];
+				const testItem = this.controller.createTestItem(toKebabCase(scenario.title), scenario.title, featureUri);
+				this.scenarioData.set(testItem, scenario);
+				scenarioItems.push(testItem);
+			}
+			if (parsedFeature.scenarioOutlines.length > 0) {
+				for (let soIdx = 0; soIdx < parsedFeature.scenarioOutlines.length; soIdx++) {
+					const scenarioOutline = parsedFeature.scenarioOutlines[soIdx];
+					if (scenarioOutline.exampleScenarios.length > 0) {
+						const scenario = scenarioOutline.exampleScenarios[0];
+						scenario.lineNumber = scenarioOutline.lineNumber;
+						const testItem = this.controller.createTestItem(
+							toKebabCase(scenarioOutline.title),
+							scenarioOutline.title,
+							featureUri
+						);
+						this.scenarioData.set(testItem, scenario);
+						scenarioItems.push(testItem);
 					}
 				}
-				testFeature.children.replace(scenarioItems);
 			}
+			testFeature.children.replace(scenarioItems);
 		}
 	}
 
@@ -159,6 +132,7 @@ export default class CucumberTestFeatures {
 	public async runCodeLenseTests(
 		testFeatureSteps: Array<TestFeatureStep>,
 		cancellationToken: vscode.CancellationToken,
+		profileName: string,
 		debug: boolean = false
 	) {
 		const include = new Array<TestItem>();
@@ -192,7 +166,7 @@ export default class CucumberTestFeatures {
 		);
 		if (include.length > 0) {
 			const request = new vscode.TestRunRequest(include);
-			await this.runTests(request, cancellationToken, debug, true);
+			await this.runTests(request, cancellationToken, profileName, debug, true);
 		}
 	}
 
@@ -203,6 +177,7 @@ export default class CucumberTestFeatures {
 	public async runTests(
 		request: vscode.TestRunRequest,
 		cancellationToken: vscode.CancellationToken,
+		profileName: string,
 		debug: boolean = false,
 		codeLense: boolean = false
 	) {
@@ -215,18 +190,20 @@ export default class CucumberTestFeatures {
 		if (request.include) {
 			const sortedItems = sortByTestLabel(request.include);
 			if (!debug && codeLense) {
-				await Promise.all(sortedItems.map(testItem => this.runTest(testItem, request, run, cancellationToken, debug)));
+				await Promise.all(
+					sortedItems.map(testItem => this.runTest(testItem, request, run, cancellationToken, profileName, debug))
+				);
 			} else {
 				const itemsLen = sortedItems.length;
 				for (let idx = 0; idx < itemsLen; idx++) {
-					await this.runTest(sortedItems[idx], request, run, cancellationToken, debug);
+					await this.runTest(sortedItems[idx], request, run, cancellationToken, profileName, debug);
 				}
 			}
 		} else if (this.testItems) {
 			const sortedItems = sortByTestLabel(this.testItems);
 			const itemsLen = sortedItems.length;
 			for (let idx = 0; idx < itemsLen; idx++) {
-				await this.runTest(sortedItems[idx], request, run, cancellationToken, debug);
+				await this.runTest(sortedItems[idx], request, run, cancellationToken, profileName, debug);
 			}
 		}
 		run.end();
@@ -253,11 +230,8 @@ export default class CucumberTestFeatures {
 		const featureUri = vscode.Uri.file(normalizePath(filePath));
 		const parsedFeature = await this.stepFileManager.getParsedFeature(featureUri);
 		if (parsedFeature && (parsedFeature.scenarios.length > 0 || parsedFeature.scenarioOutlines.length > 0)) {
-			let projectFeature = await this.getProjectFeature(parsedFeature);
-			if (projectFeature) {
-				const featureId = toKebabCase(parsedFeature.title);
-				testFeature = projectFeature.featureTestItems.find(x => x.id === featureId);
-			}
+			const featureId = toKebabCase(parsedFeature.title);
+			testFeature = this.testItems.find(x => x.id === featureId);
 		}
 		return testFeature;
 	}
@@ -274,6 +248,7 @@ export default class CucumberTestFeatures {
 		request: vscode.TestRunRequest,
 		run: vscode.TestRun,
 		cancellationToken: vscode.CancellationToken,
+		profileName: string,
 		debug: boolean
 	) {
 		// Users can hide or filter out tests from their run. If the request says
@@ -281,7 +256,7 @@ export default class CucumberTestFeatures {
 		if (request.exclude?.includes(testItem)) {
 			return;
 		}
-		await this.testRunner(testItem, run, cancellationToken, debug);
+		await this.testRunner(testItem, run, cancellationToken, profileName, debug);
 	}
 
 	/**
@@ -294,6 +269,7 @@ export default class CucumberTestFeatures {
 		testItem: vscode.TestItem,
 		run: vscode.TestRun,
 		cancellationToken: vscode.CancellationToken,
+		profileName: string,
 		debug: boolean
 	) {
 		if (testItem.children && testItem.children.size > 0) {
@@ -302,10 +278,12 @@ export default class CucumberTestFeatures {
 			const sortedChildren = sortByTestLabel(children);
 			if (debug || hasScenario === false) {
 				for (let idx = 0; idx < sortedChildren.length; idx++) {
-					await this.testRunner(sortedChildren[idx], run, cancellationToken, debug);
+					await this.testRunner(sortedChildren[idx], run, cancellationToken, profileName, debug);
 				}
 			} else {
-				await Promise.all(sortedChildren.map(childNode => this.testRunner(childNode, run, cancellationToken, debug)));
+				await Promise.all(
+					sortedChildren.map(childNode => this.testRunner(childNode, run, cancellationToken, profileName, debug))
+				);
 			}
 		} else {
 			const scenario = this.scenarioData.get(testItem);
@@ -315,25 +293,12 @@ export default class CucumberTestFeatures {
 					testItem,
 					run,
 					debug
-						? await this.cucumberTestRunner.debug(testItem.uri!.path, scenario.lineNumber, scenario.cwd, scenario.args)
-						: await this.cucumberTestRunner.run(testItem.uri!.path, scenario.lineNumber, scenario.cwd, scenario.args),
+						? await this.cucumberTestRunner.debug(testItem.uri!.path, scenario.lineNumber, profileName, scenario)
+						: await this.cucumberTestRunner.run(testItem.uri!.path, scenario.lineNumber, profileName, scenario),
 					cancellationToken
 				);
 			}
 		}
-	}
-
-	/**
-	 * Private helper used to get a projectFeature.
-	 * Will load tests if not already loaded
-	 * @param feature
-	 */
-	private async getProjectFeature(feature: ParsedFeature) {
-		if (this.projectFeatures.length === 0) {
-			await this.loadTests();
-		}
-		const projectId = toKebabCase(feature.options.projectName);
-		return this.projectFeatures.find(x => x.projectItem.id === projectId);
 	}
 
 	/**