+++ src/cucumber/step-file-manager.ts
@@ -1,32 +1,36 @@
 import * as vscode from 'vscode';
-import CucumberConfig from './cucumber-config';
-import GherkinManager, { StepInfo, IMapFeaturesResult } from '../gherkin/gherkin-manager';
+import { StepInfo } from '../gherkin/gherkin-manager';
 import { CucumberProject, FeatureStepMatch, ParsedFeature } from '../types';
 import { TextDocument } from 'vscode';
-import { normalizePath } from '../utils';
-import { project } from 'types-ramda';
 
-type CucumberProjectLocal = {
-	name: string;
-	path: string;
-	gherkin: GherkinManager;
-	currentStepText?: string;
-	findFeatureResults?: IMapFeaturesResult;
+const getDefaultProfile = (profileNames: string[]): string => {
+	if (profileNames.indexOf('default') >= 0) {
+		return 'default';
+	}
+	return profileNames[0];
 };
 
 export default class StepFileManager {
-	private cucumberConfig: CucumberConfig;
 	private cucumberProject: CucumberProject;
 
-	private projects = new Array<CucumberProjectLocal>();
-
 	constructor(project: CucumberProject) {
-		this.cucumberConfig = new CucumberConfig();
 		this.cucumberProject = project;
 	}
 
 	public get hasFeatures(): boolean {
 		return this.cucumberProject.gherkin.parsedFeatures.length > 0;
+	}
+
+	public get projectName(): string {
+		return this.cucumberProject.name;
+	}
+
+	public get projectTagPattern(): string | undefined {
+		const profileNames = Object.getOwnPropertyNames(this.cucumberProject.config);
+		const defaultProfile = getDefaultProfile(profileNames);
+
+		const profile = this.cucumberProject.config[defaultProfile];
+		return profile.tags;
 	}
 
 	/**
@@ -48,7 +52,9 @@ export default class StepFileManager {
 	};
 
 	public getParsedFeature = async (uri: vscode.Uri): Promise<ParsedFeature | undefined> => {
-		return (await this.getParsedFeatures()).find((x: { featureFile: string }) => x.featureFile === uri.fsPath);
+		return this.cucumberProject.gherkin.parsedFeatures.find(
+			(x: { featureFile: string }) => x.featureFile === uri.fsPath
+		);
 	};
 
 	/**
@@ -56,7 +62,7 @@ export default class StepFileManager {
 	 * @param stepText
 	 * @returns
 	 */
-	public getFeaturesAndScenarios = (filePath: string, stepText: string): Array<FeatureStepMatch> => {
+	public getFeaturesAndScenarios = (stepText: string): Array<FeatureStepMatch> => {
 		if (this.cucumberProject.findFeatureResults?.features) {
 			return this.cucumberProject.findFeatureResults.features.getMatchingFeaturesAndScenarios(stepText);
 		}
@@ -78,19 +84,5 @@ export default class StepFileManager {
 			profile = Object.values(this.cucumberProject.config)[0];
 		}
 		await this.cucumberProject.gherkin.loadFeatures(profile.paths);
-
-		// let project = this.projects.find(x => x.path === cucumberRoot);
-		// if (!project) {
-		// 	const nameIdx = cucumberRoot.lastIndexOf('/');
-		// 	const projectName = cucumberRoot.substring(nameIdx + 1);
-		// 	project = {
-		// 		name: projectName,
-		// 		path: cucumberRoot,
-		// 		gherkin: new GherkinManager(projectName, cucumberRoot)
-		// 	} as CucumberProjectLocal;
-		// 	const config = await this.cucumberConfig.loadCucumberConfig(cucumberRoot);
-		// 	await project.gherkin.loadFeatures(config.paths);
-		// 	this.projects.push(project);
-		// }
 	};
 }